renamed custom_model_demo
