+++ watertap/examples/custom_model_demo/demo_simple_filter.py
@@ -6,8 +6,8 @@ from idaes.core.util.model_statistics import degrees_of_freedom
 import idaes.core.util.scaling as iscale
 from idaes.core.solvers import get_solver
 
-import watertap.examples.simple_demo.simple_prop_pack as props
-from watertap.examples.simple_demo.simple_filter import Filtration
+import watertap.examples.custom_model_demo.simple_prop_pack as props
+from watertap.examples.custom_model_demo.simple_filter import Filtration
 
 
 def main():
