+++ watertap/examples/custom_model_demo/tests/test_demo_simple_filter.py
@@ -19,7 +19,7 @@ from idaes.core.util import get_solver
 from idaes.core.util.model_statistics import degrees_of_freedom
 import idaes.core.util.scaling as iscale
 
-from watertap.examples.simple_demo.demo_simple_prop_pack import main
+from watertap.examples.custom_model_demo.demo_simple_filter import main
 
 solver = get_solver()
 
@@ -28,23 +28,12 @@ solver = get_solver()
 def test_demo_simple_prop_pack():
     m = main()
 
-    assert pytest.approx(1.471, rel=1e-3) == value(
-        m.fs.stream[0].flow_mass_phase_comp["Liq", "H2O"]
+    assert pytest.approx(0.97, rel=1e-3) == value(
+        m.fs.filter.properties_out[0].flow_mass_phase_comp["Liq", "H2O"]
     )
-    assert pytest.approx(0.07746, rel=1e-3) == value(
-        m.fs.stream[0].flow_mass_phase_comp["Liq", "NaCl"]
+    assert pytest.approx(3.395e-2, rel=1e-3) == value(
+        m.fs.filter.properties_out[0].flow_mass_phase_comp["Liq", "NaCl"]
     )
-    assert pytest.approx(1.239e-4, rel=1e-3) == value(
-        m.fs.stream[0].flow_mass_phase_comp["Liq", "TSS"]
-    )
-    assert pytest.approx(298.15, rel=1e-3) == value(m.fs.stream[0].temperature)
-    assert pytest.approx(1.01325e5, rel=1e-3) == value(m.fs.stream[0].pressure)
-    assert pytest.approx(0.9499, rel=1e-3) == value(
-        m.fs.stream[0].mass_frac_phase_comp["Liq", "H2O"]
-    )
-    assert pytest.approx(51.64, rel=1e-3) == value(
-        m.fs.stream[0].conc_mass_phase_comp["Liq", "NaCl"]
-    )
-    assert pytest.approx(1032.8, rel=1e-3) == value(
-        m.fs.stream[0].dens_mass_phase["Liq"]
+    assert pytest.approx(1.200e-5, rel=1e-3) == value(
+        m.fs.filter.properties_out[0].flow_mass_phase_comp["Liq", "TSS"]
     )
